add list-instant-create-button helper
